dynamodbのmoduleを作成し実装位置を移動
