+++ wordwolf-backend/src/dynamodb/dynamodb.module.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { DynamoDB } from "./dynamodb";
+
+@Module({
+  providers: [DynamoDB],
+  exports: [DynamoDB],
+})
+export class DynamoDBModule {}
