+++ wordwolf-backend/src/words/words.service.spec.ts
@@ -2,13 +2,17 @@ import { ConfigModule } from "@nestjs/config";
 import { Test, TestingModule } from "@nestjs/testing";
 import { WordsService } from "./words.service";
 import configuration from "@/config/configuration";
+import { DynamoDBModule } from "@/dynamodb/dynamodb.module";
 
 describe("WordsService", () => {
   let service: WordsService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [ConfigModule.forRoot({ load: [configuration], envFilePath: [".env.local", ".env"] })],
+      imports: [
+        ConfigModule.forRoot({ load: [configuration], envFilePath: [".env.local", ".env"], isGlobal: true }),
+        DynamoDBModule,
+      ],
       providers: [WordsService],
     }).compile();
