+++ wordwolf-backend/src/words/words.service.ts
@@ -1,29 +1,11 @@
 import { Injectable } from "@nestjs/common";
-import { ConfigService } from "@nestjs/config";
-import { DynamoDBClient, DynamoDBClientConfig, ListTablesCommand } from "@aws-sdk/client-dynamodb";
+import { DynamoDB } from "@/dynamodb/dynamodb";
 
 @Injectable()
 export class WordsService {
-  private dynamodbClient: DynamoDBClient;
-
-  constructor(configService: ConfigService) {
-    const dynamodbClientConfig: DynamoDBClientConfig = {
-      region: configService.get<string>("dynamodb.region", ""),
-    };
-
-    const endpoint = configService.get<string | undefined>("dynamodb.endpoint");
-    if (endpoint != undefined) {
-      // endpointを指定するのはローカルで動作させるときだけ。
-      // それ以外のときは指定せず、AWSのエンドポイントを使う。
-      dynamodbClientConfig.endpoint = endpoint;
-    }
-
-    this.dynamodbClient = new DynamoDBClient(dynamodbClientConfig);
-  }
+  constructor(private readonly dynamoDB: DynamoDB) {}
 
   async listTables(): Promise<string> {
-    const command = new ListTablesCommand({});
-    const result = await this.dynamodbClient.send(command);
-    return JSON.stringify(result);
+    return this.dynamoDB.listTables();
   }
 }
